% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_Schroeder_et_al_2016_OEC.R
\docType{data}
\name{df_Schroeder_et_al_2016_OEC}
\alias{df_Schroeder_et_al_2016_OEC}
\title{Example dataset for random effects (predator individual)}
\format{
A data frame with 686 rows and 8 variables:
\describe{
\item{N0}{Number of initial prey}
\item{NE}{Number of eaten prey}
\item{Time}{Duration (h): 2 min}
\item{ID}{Predator individual ID}
\item{Size}{Predator individual size (mm)}
\item{Predator}{Predator species}
\item{Prey}{Prey species}
\item{Trial.time}{Trials performed in the morning or evening}
}
}
\source{
Schröder A., Kalinkat G. & Arlinghaus R. (2016).
\emph{Individual variation in functional response parameters is explained by body size but not by behavioural types in a poeciliid fish.}
Oecologia, 88:1670–1683.
https://doi.org/10.1007/s00442-016-3701-7

Kalinkat G., Schröder A. & Arlinghaus R. (2025).
\emph{Individual variation in functional response parameters is explained by body size but not by behavioural types in a poeciliid fish.}
Figshare.
https://doi.org/10.6084/m9.figshare.24665880
}
\usage{
data(df_Schroeder_et_al_2016_OEC)
}
\description{
Feeding experiment data from Schröder et al. (2016) were downloaded from
Figshare (Kalinkat et al. 2025) under CC BY 4.0. Eaten prey were not replaced during the
experiments (2 minutes). Includes data for least killifish (Heterandria formosa, 49 individuals)
feeding on nauplii (Artemia salina). Predator individuals were re-used and predator
ID was recorded for each trial. Also includes predator size.
}
\examples{
data(df_Schroeder_et_al_2016_OEC)
head(df_Schroeder_et_al_2016_OEC)
}
\keyword{datasets}
