% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_Sentis_et_al_2017_GLOBAL_CHANGE_BIOLOGY.R
\docType{data}
\name{df_Sentis_et_al_2017_GLOBAL_CHANGE_BIOLOGY}
\alias{df_Sentis_et_al_2017_GLOBAL_CHANGE_BIOLOGY}
\title{Example dataset for testing type 2 vs. type 3}
\format{
A data frame with 327 rows and 7 variables:
\describe{
\item{N0}{Number of initial prey}
\item{NE}{Number of eaten prey}
\item{Time}{Duration (h)}
\item{Predator}{Predator species}
\item{Prey}{Prey species}
\item{ID}{Predator-Temperature combination}
\item{Temperature}{Temperature treatment}
}
}
\source{
Sentis A., Gemard C., Jaugeon B., and Boukal D. S. (2017).
\emph{Predator diversity and environmental change modify the strengths of trophic and nontrophic interactions.}
Global Change Biology, 23: 2629-2640.
https://doi.org/10.1111/gcb.13560

Uiterwaal S. F., Lagerstrom I. T., Lyon S. R., and DeLong, J. P. (2022).
\emph{FoRAGE Database: A Compilation of Functional Responses for Consumers and Parasitoids.}
Ecology 103(7): e3706.
https://doi.org/10.1002/ecy.3706

FoRAGE database V5 (2024).
https://doi.org/10.5063/F1RX99KB
}
\usage{
data(df_Sentis_et_al_2017_GLOBAL_CHANGE_BIOLOGY)
}
\description{
Feeding experiment data from Sentis et al. (2017) downloaded from the
FoRAGE database (Uiterwaal et al. 2022). Eaten prey were not replaced during the
experiment. Includes data for three aquatic insect larvae predators feeding on
Daphnia prey in two temperature treatments.
}
\examples{
data(df_Sentis_et_al_2017_GLOBAL_CHANGE_BIOLOGY)
head(df_Sentis_et_al_2017_GLOBAL_CHANGE_BIOLOGY)
}
\keyword{datasets}
