% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MS_Type3H_dyn_code.R
\name{MS_Type3H_dyn_code}
\alias{MS_Type3H_dyn_code}
\title{Type 3 multi-species FR without replacement}
\usage{
MS_Type3H_dyn_code
}
\value{
Number of eaten prey
}
\description{
Contains Stan function of the same name as character string. Uses numerical
solution of the 2-prey ODE
\deqn{\frac{dN_1}{dt}=-\frac{b_1N_1^{1+q}}{1+b_1h_1N_1^{1+q}+b_2h_2N_2^{1+q}}P}{dN1/dt=-b1N1^(1+q) P/(1+b1h1N1^(1+q)+b2h2N2^(1+q)))}
\deqn{\frac{dN_2}{dt}=-\frac{b_2N_2^{1+q}}{1+b_1h_1N_1^{1+q}+b_2h_2N_2^{1+q}}P}{dN2/dt=-b2N2^(1+q) P/(1+b1h1N1^(1+q)+b2h2N2^(1+q)))}
to compute number of eaten prey.
}
\details{
Usage in brms formula:\cr\cr
\verb{~ MS_Type3H_dyn(N0, N0.alt, ID, P, Time, b1, b2, h1, h2, q)}\cr\cr
\verb{N0    } initial number of focal prey species \cr
\verb{N0.alt} initial number of alternative prey species \cr
\verb{ID    } ID of focal species (1 or 2) \cr
\verb{P     } number of predators \cr
\verb{Time  } duration of the experiment \cr
\verb{b1    } attack coefficient for prey species 1 \cr
\verb{b2    } attack coefficient for prey species 2 \cr
\verb{h1    } handling time for prey species 1 \cr
\verb{h2    } handling time for prey species 2 \cr
\verb{q     } attack rate exponent \cr

Requires the data to be in a specific "long" format. Use the function
\verb{convert_2sp_to_long()} to transform from a dataframe with columns
N01, N02, NE1, NE2
}
