% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MS_TypeZ_dyn_code.R
\name{MS_TypeZ_dyn_code}
\alias{MS_TypeZ_dyn_code}
\title{Generalized switching FR without replacement}
\usage{
MS_TypeZ_dyn_code
}
\value{
Number of eaten prey
}
\description{
Contains Stan function of the same name as character string. Uses numerical
solution of the 2-prey ODE
\deqn{\frac{dN_1}{dt}=-\frac{w_1b_1N_1^{(1+q+r)}}{(w_1N_1^r+(1-w_1)N_2^r)+w_1b_1h_1N_1^{(1+q+r)}+(1-w_1)b_2h_2N_2^{(1+q+r)}}P}{dN1/dt=-w1b1N1^(1+q+r) P/((w1N1^r+(1+w1)N2^r)+w1b1h1N1^(1+q+r)+(1-w1)b2h2N2^(1+q+r)))}
\deqn{\frac{dN_2}{dt}=-\frac{(1-w_1)b_2N_2^{(1+q+r)}}{(w_1N_1^r+(1-w_1)N_2^r)+w_1b_1h_1N_1^{(1+q+r)}+(1-w_1)b_2h_2N_2^{(1+q+r)}}P}{dN2/dt=-w1b2N2^(1+q+r) P/((w1N1^r+(1+w1)N2^r)+w1b1h1N1^(1+q+r)+(1-w1)b2h2N2^(1+q+r)))}
to compute number of eaten prey.
}
\details{
Usage in brms formula:\cr\cr
\verb{~ MS_TypeZ_dyn(N0, N0.alt, ID, P, Time, b1, b2, h1, h2, w1, q r)}\cr\cr
\verb{N0    } initial number of focal prey species \cr
\verb{N0.alt} initial number of alternative prey species \cr
\verb{ID    } ID of focal species (1 or 2) \cr
\verb{P     } number of predators \cr
\verb{Time  } duration of the experiment \cr
\verb{b1    } attack rate for prey species 1 \cr
\verb{b2    } attack rate for prey species 2 \cr
\verb{h1    } handling time for prey species 1 \cr
\verb{h2    } handling time for prey species 2 \cr
\verb{w1    } preference weight for species 1 in multi-prey \cr
\verb{q     } attack rate exponent \cr
\verb{r     } additional exponent in multi-prey only \cr

Requires the data to be in a specific "long" format. Use the function
\verb{convert_2sp_to_long()} to transform from a dataframe with columns
N01, N02, NE1, NE2
}
