% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Type1_dyn.R
\docType{data}
\name{Type1_dyn}
\alias{Type1_dyn}
\title{Type 1 FR with prey depletion}
\format{
An object of class \code{character} of length 1.
}
\usage{
Type1_dyn
}
\value{
Number of eaten prey
}
\description{
Contains Stan function of the same name as character string. Uses analytical
solution to compute number of eaten prey.
}
\details{
Usage in brms formula:\cr\cr
\verb{~ Type1_dyn(N, a, P, Time)}\cr\cr
\verb{N    } initial number of prey \cr
\verb{a    } attack rate \cr
\verb{P    } number of predators \cr
\verb{Time } duration of the experiment
}
\keyword{datasets}
