% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Type1_dyn_code.R
\name{Type1_dyn_code}
\alias{Type1_dyn_code}
\title{Type 1 FR with prey depletion}
\usage{
Type1_dyn_code
}
\value{
Number of eaten prey
}
\description{
Contains Stan function of the same name as character string. Uses analytical
solution (exponential function) of the ODE
\deqn{\frac{dN}{dt}=-aNP}{dN/dt=-aNP}
to compute number of eaten prey.
}
\details{
Usage in brms formula:\cr\cr
\verb{~ Type1_dyn(N, P, Time, a)}\cr\cr
\verb{N    } initial number of prey \cr
\verb{P    } number of predators \cr
\verb{Time } duration of the experiment \cr
\verb{a    } attack rate
}
