% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Type1_fix.R
\docType{data}
\name{Type1_fix}
\alias{Type1_fix}
\title{Type 1 FR with prey replacement}
\format{
An object of class \code{character} of length 1.
}
\usage{
Type1_fix
}
\value{
Number of eaten prey
}
\description{
Contains Stan function of the same name as character string.
Number of eaten prey:
\deqn{N_E=aNPT}{N_E=aNPT}
}
\details{
Usage in brms formula:\cr\cr
\verb{~ Type1_fix(N, a, P, Time)}\cr\cr
\verb{N    } initial number of prey \cr
\verb{a    } attack rate \cr
\verb{P    } number of predators \cr
\verb{Time } duration of the experiment
}
\keyword{datasets}
