% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Type2CM_dyn_code.R
\docType{data}
\name{Type2CM_dyn_code}
\alias{Type2CM_dyn_code}
\title{Functional response models with predator interference}
\format{
An object of class \code{character} of length 1.
}
\usage{
Type2CM_dyn_code
}
\value{
Number of eaten prey
}
\description{
Contains Stan function of the same name as character string. Uses analytical
solution of the \bold{Crowley-Martin} model
\deqn{\frac{dN}{dt}=-\frac{aN}{(1+ahN)(1+c(P-1))}P}{dN/dt=-aNP/((1+ahN)(1+c(P-1)))}
to compute number of eaten prey. Rogers random predator equation with LambertW
function is used with modified attack rates and handling times. Predator
interference affects attack rates and handling times, both.
}
\details{
Usage in brms formula:\cr\cr
\verb{~ Type2CM_dyn(N, P, Time, a, h, c)}\cr\cr
\verb{N    } initial number of prey \cr
\verb{P    } number of predators \cr
\verb{Time } duration of the experiment \cr
\verb{a    } attack rate for P=1\cr
\verb{h    } handling time \cr
\verb{c    } predator interference coefficient
}
\keyword{datasets}
