% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Type2H_dyn_code.R
\docType{data}
\name{Type2H_dyn_code}
\alias{Type2H_dyn_code}
\title{Type 2 FR (Holling) with prey depletion}
\format{
An object of class \code{character} of length 1.
}
\usage{
Type2H_dyn_code
}
\value{
Number of eaten prey
}
\description{
Contains Stan function of the same name as character string. Uses analytical
solution (Rogers random predator equation with LambertW function) of the ODE
\deqn{\frac{dN}{dt}=-\frac{aN}{1+ahN}P}{dN/dt=-aNP/(1+ahN)}
to compute number of eaten prey.
}
\details{
Usage in brms formula:\cr\cr
\verb{~ Type2H_dyn(N, P, Time, a, h)}\cr\cr
\verb{N    } initial number of prey \cr
\verb{P    } number of predators \cr
\verb{Time } duration of the experiment \cr
\verb{a    } attack rate \cr
\verb{h    } handling time
}
\keyword{datasets}
