% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Type2H_fix_code.R
\docType{data}
\name{Type2H_fix_code}
\alias{Type2H_fix_code}
\title{Type 2 FR (Holling) with prey replacement}
\format{
An object of class \code{character} of length 1.
}
\usage{
Type2H_fix_code
}
\value{
Number of eaten prey
}
\description{
Contains Stan function of the same name as character string.
Number of eaten prey:
\deqn{N_E=\frac{aN}{1+ahN}PT}{N_E=aNPT/(1+ahN)}
}
\details{
Usage in brms formula:\cr\cr
\verb{~ Type2H_fix(N, a, h, P, Time)}\cr\cr
\verb{N    } initial number of prey \cr
\verb{a    } attack rate \cr
\verb{h    } handling time \cr
\verb{P    } number of predators \cr
\verb{Time } duration of the experiment
}
\keyword{datasets}
