% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Type3GenH_fix.R
\docType{data}
\name{Type3GenH_fix}
\alias{Type3GenH_fix}
\title{Generalized type 3 FR (Holling) with prey replacement}
\format{
An object of class \code{character} of length 1.
}
\usage{
Type3GenH_fix
}
\value{
Number of eaten prey
}
\description{
Contains Stan function of the same name as character string.
Number of eaten prey:
\deqn{N_E=\frac{bN^{1+q}}{1+bhN^{1+q}}PT}{N_E=bN^(1+q) PT/(1+bhN^(1+q))}
}
\details{
Usage in brms formula:\cr\cr
\verb{~ Type3GenH_fix(N, b, h, q, P, Time)}\cr\cr
\verb{N    } initial number of prey \cr
\verb{b    } attack coefficient \cr
\verb{q    } attack exponent \cr
\verb{h    } handling time \cr
\verb{P    } number of predators \cr
\verb{Time } duration of the experiment
}
\keyword{datasets}
