% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Type3GenH_mort_dyn_code.R
\docType{data}
\name{Type3GenH_mort_dyn_code}
\alias{Type3GenH_mort_dyn_code}
\title{Functional response models with prey mortality}
\format{
An object of class \code{character} of length 1.
}
\usage{
Type3GenH_mort_dyn_code
}
\value{
Number of eaten prey
}
\description{
Contains Stan function of the same name as character string. Uses numerical
solution of the generalized type 3 ODE with an additional mortality term
\deqn{\frac{dN}{dt}=-\frac{bN^{1+q}}{1+bhN^{1+q}}P-mN}{dN/dt=-bN^(1+q) P/(1+bhN^(1+q))-mN}
to compute number of eaten / dead prey. It can compute predictions for feeding
trials (observations with \verb{P>0}) and also control treatments (\verb{P=0}),
for which the ODE reduces to
\deqn{\frac{dN}{dt}=-mN}{dN/dt=-mN}
The exponent \verb{q} can be fixed for fitting type 2 responses (\verb{q=0}) or
type 3 responses (\verb{q=1}), which both do not have an analytical solution
with additional and prey mortality have to be predicted using the ODE.
}
\details{
Usage in brms formula:\cr\cr
\verb{~ Type3GenH_dyn(N, P, Time, b, h, q, m)}\cr\cr
\verb{N    } initial number of prey \cr
\verb{P    } number of predators \cr
\verb{Time } duration of the experiment \cr
\verb{b    } attack coefficient \cr
\verb{q    } attack exponent \cr
\verb{h    } handling time \cr
\verb{m    } mortality rate

Type 2 functional response: \cr\cr
\verb{~ Type3GenH_dyn(N, P, Time, a, h, 0, m)}\cr\cr
Type 3 functional response: \cr\cr
\verb{~ Type3GenH_dyn(N, P, Time, b, h, 1, m)}
}
\keyword{datasets}
