% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Type3H_dyn_code.R
\docType{data}
\name{Type3H_dyn_code}
\alias{Type3H_dyn_code}
\title{Type 3 FR (Holling) with prey depletion}
\format{
An object of class \code{character} of length 1.
}
\usage{
Type3H_dyn_code
}
\value{
Number of eaten prey
}
\description{
Contains Stan function of the same name as character string. Uses analytical
solution (quadratic equation) of the ODE
\deqn{\frac{dN}{dt}=-\frac{bN^2}{1+bhN^2}P}{dN/dt=-bN^2 P/(1+bhN^2)}
to compute number of eaten prey.
}
\details{
Usage in brms formula:\cr\cr
\verb{~ Type3H_dyn(N, P, Time, b, h)}\cr\cr
\verb{N    } initial number of prey \cr
\verb{P    } number of predators \cr
\verb{Time } duration of the experiment \cr
\verb{b    } attack coefficient \cr
\verb{h    } handling time
}
\keyword{datasets}
