% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Type3H_fix_code.R
\docType{data}
\name{Type3H_fix_code}
\alias{Type3H_fix_code}
\title{Type 3 FR (Holling) with prey replacement}
\format{
An object of class \code{character} of length 1.
}
\usage{
Type3H_fix_code
}
\value{
Number of eaten prey
}
\description{
Contains Stan function of the same name as character string.
Number of eaten prey:
\deqn{N_E=\frac{bN^2}{1+bhN^2}PT}{N_E=bN^2 PT/(1+bhN^2)}
}
\details{
Usage in brms formula:\cr\cr
\verb{~ Type3H_fix(N, P, Time, b, h)}\cr\cr
\verb{N    } initial number of prey \cr
\verb{P    } number of predators \cr
\verb{Time } duration of the experiment \cr
\verb{b    } attack coefficient \cr
\verb{h    } handling time
}
\keyword{datasets}
