% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TypeGenHV_dyn_code.R
\docType{data}
\name{TypeGenHV_dyn_code}
\alias{TypeGenHV_dyn_code}
\title{Functional response models with predator interference}
\format{
An object of class \code{character} of length 1.
}
\usage{
TypeGenHV_dyn_code
}
\value{
Number of eaten prey
}
\description{
Contains Stan function of the same name as character string. Uses numerical
solution of the \bold{(generalized) Hassell-Varley} model
\deqn{\frac{dN}{dt}=-\frac{bN^{1+q}}{P^c+bhN^{1+q}}P}{dN/dt=-bN^(1+q)P/(P^c++bhN^{1+q})}
to compute number of eaten prey. Predator interference affects attack rates only.
}
\details{
The exponent \verb{q} can be fixed for fitting type 2 responses (\verb{q=0}) or
type 3 responses (\verb{q=1}).

Usage in brms formula:\cr\cr
\verb{~ TypeGenHV_dyn(N, P, Time, b, h, q, c)}\cr\cr
\verb{N    } initial number of prey \cr
\verb{P    } number of predators \cr
\verb{Time } duration of the experiment \cr
\verb{b    } attack coefficient for P=1 \cr
\verb{q    } attack exponent \cr
\verb{h    } handling time \cr
\verb{c    } predator interference coefficient

Type 2 functional response: \cr\cr
\verb{~ TypeGenHV_dyn(N, P, Time, a, h, 0, c)}\cr\cr
Type 3 functional response: \cr\cr
\verb{~ TypeGenHV_dyn(N, P, Time, b, h, 1, c)}
}
\keyword{datasets}
