% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_2sp_to_long.R
\name{convert_2sp_to_long}
\alias{convert_2sp_to_long}
\title{Convert 2-prey data to long format}
\usage{
convert_2sp_to_long(df)
}
\arguments{
\item{df}{data frame with first four columns named: N01, N02, NE1, NE2}
}
\value{
The transformed data frame
}
\description{
Convert 2-prey data to long format
}
