% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_Cuthbert_et_al_2020_ECOL_EVOL.R
\docType{data}
\name{df_Cuthbert_et_al_2020_ECOL_EVOL}
\alias{df_Cuthbert_et_al_2020_ECOL_EVOL}
\title{Example dataset for categorical predictors}
\format{
A data frame with 358 rows and 7 variables:
\describe{
\item{N0}{Number of initial prey}
\item{NE}{Number of eaten prey}
\item{Time}{Duration (h)}
\item{Predator}{Predator species}
\item{Prey}{Prey species}
\item{PredSize}{Predator size class}
\item{PreySize}{Prey size class}
}
}
\source{
Cuthbert R. N., Wassermann R. J., Dalu T., Kaiser H., Weyl O. L. F., Dick J. T. A., Sentis A., McCoy M. W., & Alexander M.E. (2020a).
\emph{Influence of intra- and interspecific variation in predator–prey body size ratios on trophic interaction strengths.}
Ecology and Evolution, 10:5946–5962.
https://doi.org/10.1002/ece3.6332

Cuthbert R. N., Wassermann R. J., Dalu T., Kaiser H., Weyl O. L. F., Dick J. T. A., Sentis A., McCoy M. W., & Alexander M.E. (2020b).
\emph{Influence of intra- and interspecific variation in predator–prey body size ratios on trophic interaction strengths.}
Dryad Digital Repository.
https://doi.org/10.5061/dryad.7m0cfxppt
}
\usage{
data(df_Cuthbert_et_al_2020_ECOL_EVOL)
}
\description{
Feeding experiment data from Cuthbert et al. (2020a) were downloaded from
Dryad (Cuthbert et al. 2020b). Eaten prey were not replaced during the
experiment. Includes data for two fish species (largemouth bass and bluegill)
feeding on tilapia. Both predator and prey were categorized in three size classes,
each, with a full factorial treatment.
}
\examples{
data(df_Cuthbert_et_al_2020_ECOL_EVOL)
head(df_Cuthbert_et_al_2020_ECOL_EVOL)
}
\keyword{datasets}
