% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MS_Type2H_dyn_code.R
\name{MS_Type2H_dyn_code}
\alias{MS_Type2H_dyn_code}
\title{Type 2 multi-species FR without replament}
\usage{
MS_Type2H_dyn_code
}
\value{
Number of eaten prey
}
\description{
Contains Stan function of the same name as character string. Uses numerical
solution of the 2-prey ODE
\deqn{\frac{dN_1}{dt}=-\frac{a_1N_1}{1+a_1h_1N_1+a_2h_2N_2}P}{dN1/dt=-a1N1 P/(1+a1h1N1+a2h2N2))}
\deqn{\frac{dN_2}{dt}=-\frac{a_2N_2}{1+a_1h_1N_1+a_2h_2N_2}P}{dN2/dt=-a2N2 P/(1+a1h1N1+a2h2N2))}
to compute number of eaten prey.
}
\details{
Usage in brms formula:\cr\cr
\verb{~ MS_Type2H_dyn(N0, N0.alt, ID, P, Time, a1, a2, h1, h2)}\cr\cr
\verb{N0    } initial number of focal prey species \cr
\verb{N0.alt} initial number of alternative prey species \cr
\verb{ID    } ID of focal species (1 or 2) \cr
\verb{P     } number of predators \cr
\verb{Time  } duration of the experiment \cr
\verb{a1    } attack rate for prey species 1 \cr
\verb{a2    } attack rate for prey species 2 \cr
\verb{h1    } handling time for prey species 1 \cr
\verb{h2    } handling time for prey species 2 \cr

Requires the data to be in a specific "long" format. Use the function
\verb{convert_2sp_to_long()} to transform from a dataframe with columns
N01, N02, NE1, NE2
}
