% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Type2HV_dyn_code.R
\docType{data}
\name{Type2HV_dyn_code}
\alias{Type2HV_dyn_code}
\title{Functional response models with predator interference}
\format{
An object of class \code{character} of length 1.
}
\usage{
Type2HV_dyn_code
}
\value{
Number of eaten prey
}
\description{
Contains Stan function of the same name as character string. Uses analytical
solution of the \bold{Hassell-Varley} model
\deqn{\frac{dN}{dt}=-\frac{aN}{P^c+ahN}P}{dN/dt=-aNP/(P^c+ahN)}
to compute number of eaten prey. Rogers random predator equation with LambertW
function is used with modified attack rates. Predator interference affects attack
rates only.
}
\details{
Usage in brms formula:\cr\cr
\verb{~ Type2HV_dyn(N, P, Time, a, h, c)}\cr\cr
\verb{N    } initial number of prey \cr
\verb{P    } number of predators \cr
\verb{Time } duration of the experiment \cr
\verb{a    } attack rate for P=1 \cr
\verb{h    } handling time \cr
\verb{c    } predator interference coefficient
}
\keyword{datasets}
